Enhancement (commit a65cd0f, f3e4d9c): disable the introduced mechanism for debug purposes. The introduced mechanism disturbs the usage of gdb/cuda-gdb as the ANNarchyCoreXX.so library is always deleted/moved. Therefore I decided to disable this mechanism if ANNarchy is in debug mode (either by --debug on command line or by compile() argument).

+++ ANNarchy/core/NetworkManager.py
@@ -182,7 +182,7 @@ class NetworkManager(object):
         for mon in self._network[0]['monitors']:
             mon._clear()
 
-        if self._network[0]['directory'] != None:
+        if self._network[0]['directory'] != None and not Global.config["debug"]:
             # Removes the library used in last running instance
             if os.path.isfile(self._network[0]['directory']+'/ANNarchyCore' + str(0) + '.so'):
                 os.remove(self._network[0]['directory']+'/ANNarchyCore' + str(0) + '.so')
